requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.0; python_version >= '3.7'
backports.zoneinfo==0.2.1; python_version < '3.9'
brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12; python_version >= '3'
cryptography==36.0.2
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
dj-rest-auth[with_social]==2.2.4
django-allauth==0.50.0
django-cors-headers==3.11.0
django==4.0.3
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3; python_version >= '3'
oauthlib==3.2.0; python_version >= '3.6'
psycopg2-binary==2.9.3
pycparser==2.21
pyjwt[crypto]==2.3.0; python_version >= '3.6'
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
requests==2.27.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
setuptools==62.1.0; python_version >= '3.7'
sqlparse==0.4.2; python_version >= '3.5'
urllib3==1.26.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
whitenoise==6.0.0
